move it because baz says
